modified the algorithm

+++ src/AvgCalc.java
@@ -1,45 +1,45 @@
 public class AvgCalc {
-    private Double ctAvg;
-    private Double tatAvg;
-    private Double wtAvg;
-    private Double rtAvg;
+    private float ctAvg;
+    private float tatAvg;
+    private float wtAvg;
+    private float rtAvg;
 
     public AvgCalc() {
-        this.ctAvg = 0.0;
-        this.tatAvg = 0.0;
-        this.wtAvg = 0.0;
-        this.rtAvg = 0.0;
+        this.ctAvg = 0;
+        this.tatAvg = 0;
+        this.wtAvg = 0;
+        this.rtAvg = 0;
     }
 
-    public Double getCt() {
+    public float getCt() {
         return ctAvg;
     }
 
-    public void setCt(Double ctAvg) {
+    public void setCt(float ctAvg) {
         this.ctAvg = ctAvg;
     }
 
-    public Double getTat() {
+    public float getTat() {
         return tatAvg;
     }
 
-    public void setTat(Double tatAvg) {
+    public void setTat(float tatAvg) {
         this.tatAvg = tatAvg;
     }
 
-    public Double getWt() {
+    public float getWt() {
         return wtAvg;
     }
 
-    public void setWt(Double wt) {
+    public void setWt(float wt) {
         this.wtAvg = wt;
     }
 
-    public Double getRt() {
+    public float getRt() {
         return rtAvg;
     }
 
-    public void setRt(Double rt) {
+    public void setRt(float rt) {
         this.rtAvg = rt;
     }
 }
